requirements.txt
aiohttp>=3.11.11
babel>=2.16.0
discord.py>=2.4.0
python-dotenv>=1.0.1
